Fix timestamp variable type and endpoint path

+++ Alpaca.Markets/Helpers/DateTimeHelper.cs
@@ -19,8 +19,8 @@ namespace Alpaca.Markets
             return _epoch.Add(TimeSpan.FromSeconds(linuxTimeStamp));
         }
 
-        public static Int32 GetUnixTimeMillis(DateTime dateTime) {
-            return (Int32)(dateTime.Subtract(_epoch)).TotalMilliseconds;
+        public static Int64 GetUnixTimeMillis(DateTime dateTime) {
+            return (Int64)(dateTime.Subtract(_epoch)).TotalMilliseconds;
         }
     }
 }
